♻️ refactor: 에러 핸들러 수정
