+++ weather-service/src/main/java/com/waither/weatherservice/controller/WeatherController.java
@@ -1,0 +1,4 @@
+package com.waither.weatherservice.controller;
+
+public class WeatherController {
+}
