+++ weather-service/src/main/java/com/waither/weatherservice/controller/WeatherTestController.java
@@ -9,10 +9,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.waither.weatherservice.dto.AccuweatherTestRequest;
-import com.waither.weatherservice.dto.AirTestRequest;
-import com.waither.weatherservice.dto.ForeCastTestRequest;
-import com.waither.weatherservice.dto.MsgTestRequest;
+import com.waither.weatherservice.dto.request.AccuweatherTestRequest;
+import com.waither.weatherservice.dto.request.AirTestRequest;
+import com.waither.weatherservice.dto.request.ForeCastTestRequest;
+import com.waither.weatherservice.dto.request.MsgTestRequest;
 import com.waither.weatherservice.service.WeatherService;
 
 import lombok.RequiredArgsConstructor;
