+++ weather-service/src/main/java/com/waither/weatherservice/dto/request/AccuweatherTestRequest.java
@@ -1,4 +1,4 @@
-package com.waither.weatherservice.dto;
+package com.waither.weatherservice.dto.request;
 
 public record AccuweatherTestRequest(
 	double latitude,
