+++ weather-service/src/main/java/com/waither/weatherservice/dto/request/WeatherRequest.java
@@ -1,0 +1,7 @@
+package com.waither.weatherservice.dto.request;
+
+import lombok.Builder;
+
+@Builder
+public record WeatherRequest() {
+}
