+++ weather-service/src/main/java/com/waither/weatherservice/dto/response/WeatherResponse.java
@@ -1,0 +1,7 @@
+package com.waither.weatherservice.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record WeatherResponse() {
+}
