--- weather-service/src/main/java/com/waither/weatherservice/openapi/OpenApiException.java
@@ -1,7 +1,0 @@
-package com.waither.weatherservice.openapi;
-
-public class OpenApiException extends RuntimeException {
-	public OpenApiException(String message) {
-		super(message);
-	}
-}
